refactor: 변수 네이밍, controller->service 로직 이동, Errorcode 수정
