+++ src/main/java/wanted/n/scheduler/StatisticsScheduler.java
@@ -2,7 +2,6 @@ package wanted.n.scheduler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import wanted.n.enums.DateType;
